requirements.txt
streamlit 
google-cloud-vision
google-generativeai
firebase-admin
python-dotenv
pillow
pandas
